Move the InMemory implementation of `/pkg/limits` in its own file
